refactor: close all partial TODOs (01 BUG-4, 06 finish, 10 finish)
Closes the four remaining partial TODOs by addressing each gap
identified in their respective files.

## TODO 01 BUG-4 — SequentialMember field name verification

Verified against glossarist-js Phase 1 (src/models/sequential-member.ts):
SequentialMember extends HyperedgeMember and adds NOTHING — no
is_delimiting, no delimitingCharacteristic, no characteristic.

The sequential-relations.mdx docs page incorrectly listed:
- +is_delimiting: Boolean on SequentialMember (doesn't exist)
- characteristic: YAML keys in wire format (not a modeled field)

Fixed:
- Model block: removed +is_delimiting from SequentialMember; added
  note that sequential members inherit only the MECE pair and are
  distinguished by array position
- Wire format: removed characteristic: keys from all 5 member entries
- Replaced the misleading "characteristic field" prose with "member
  array order IS significant" (the actual distinguishing property)

## TODO 06 finish — eliminate remaining any types

Remaining any-typed code in src/:
1. Nav.astro — pagefind: any + two (r: any) arrows
   → Defined PagefindInstance + PagefindSearchResult interfaces inline
   → Replaced runtimeImport return type from Promise<any> to typed
   → Removed (r: any) annotations — TypeScript infers from the typed
     PagefindSearchResult interface

2. ReleaseDownloader.astro — (data.assets as any[])
   → Defined inline GitHub release type at the fetch call site
   → Cast the res.json() result to the typed shape

Result: zero any-types in production code paths.

## TODO 10 finish — extract i18n types to dedicated module

Created src/i18n/types.ts with:
- Locale union type
- LocaleMeta interface
- TranslationKey union (22 keys)
- TranslationSet = Record<TranslationKey, string>

translations.ts now imports types from ./types instead of defining
them inline. Separation of concerns: types are importable without
pulling the full ~4KB translation data.

## Additional fix — dark-mode CSS cleanup

hyperedge-partitive.svg's dark-mode media query referenced
.tooth-delimit (not used by this SVG). Removed to prevent the
test invariant from flagging it.

## Build + tests

- npm run build: 103 pages built
- npm test: 611 tests pass (unchanged — all fixes are type-level)
